% Auto-generated: do not edit by hand
\name{''MuiButton}

\alias{''MuiButton}

\title{MuiButton component}

\description{

}

\usage{
''MuiButton(children=NULL, id=NULL, classes=NULL, color=NULL,
component=NULL, debounceWait=NULL, disableElevation=NULL,
disableFocusRipple=NULL, disableRipple=NULL, disabled=NULL,
endIcon=NULL, fullWidth=NULL, href=NULL, key=NULL,
nClicks=NULL, size=NULL, startIcon=NULL, variant=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. 组件型，按钮内嵌元素}

\item{id}{Character. 组件唯一id}

\item{classes}{Named list. 覆盖或扩展应用于组件的样式}

\item{color}{A value equal to: 'inherit', 'primary', 'secondary', 'success', 'error', 'info', 'warning', 'string'. 按钮颜色}

\item{component}{按钮组件}

\item{debounceWait}{Numeric. 防抖等待时间}

\item{disableElevation}{Logical. 是否禁用按钮阴影}

\item{disableFocusRipple}{Logical. 是否禁用按钮聚焦时的涟漪效果}

\item{disableRipple}{Logical. 是否禁用按钮涟漪效果}

\item{disabled}{Logical. 是否禁用按钮}

\item{endIcon}{A list of or a singular dash component, string or number. 按钮尾部图标}

\item{fullWidth}{Logical. 是否自动扩展按钮宽度}

\item{href}{Character. 按钮链接}

\item{key}{Character. 组件key}

\item{nClicks}{Numeric. 按钮累计点击次数，用于监听按钮点击行为
默认值：`0`}

\item{size}{A value equal to: 'small', 'medium', 'large', 'string'. 按钮大小}

\item{startIcon}{A list of or a singular dash component, string or number. 按钮头部图标}

\item{variant}{A value equal to: 'outlined', 'contained', 'text', 'string'. 组件样式}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

