% Auto-generated: do not edit by hand
\name{''MuiTextField}

\alias{''MuiTextField}

\title{MuiTextField component}

\description{

}

\usage{
''MuiTextField(id=NULL, FormHelperTextProps=NULL, InputLabelProps=NULL,
InputProps=NULL, SelectProps=NULL, autoComplete=NULL,
autoFocus=NULL, classes=NULL, color=NULL, defaultValue=NULL,
disabled=NULL, error=NULL, fullWidth=NULL, helperText=NULL,
inputProps=NULL, inputRef=NULL, label=NULL, margin=NULL,
maxRows=NULL, minRows=NULL, multiline=NULL, name=NULL,
onChange=NULL, placeholder=NULL, required=NULL, rows=NULL,
select=NULL, size=NULL, sx=NULL, type=NULL, value=NULL,
variant=NULL)
}

\arguments{
\item{id}{Character. 输入元素的ID}

\item{FormHelperTextProps}{Named list. 应用于FormHelperText元素的属性}

\item{InputLabelProps}{Named list. 应用于InputLabel元素的属性}

\item{InputProps}{Named list. 应用于输入组件的属性}

\item{SelectProps}{Named list. 应用于选择框元素的属性}

\item{autoComplete}{Character. 自动填充的类型}

\item{autoFocus}{Logical. 是否自动聚焦}

\item{classes}{Named list. 覆盖或扩展应用于组件的样式}

\item{color}{A value equal to: 'primary', 'secondary', 'error', 'info', 'success', 'warning', 'string'. 组件的颜色}

\item{defaultValue}{Logical | numeric | character | named list | unnamed list. 默认值}

\item{disabled}{Logical. 是否禁用组件}

\item{error}{Logical. 是否显示错误状态}

\item{fullWidth}{Logical. 是否占满父容器的宽度}

\item{helperText}{A list of or a singular dash component, string or number. 帮助文本内容}

\item{inputProps}{Named list. 应用于输入元素的属性}

\item{inputRef}{Lists containing elements 'current'.
those elements have the following types:
  - current (logical | numeric | character | named list | unnamed list; optional). 传递给输入元素的ref}

\item{label}{A list of or a singular dash component, string or number. 标签内容}

\item{margin}{A value equal to: 'dense', 'none', 'normal'. 垂直间距调整}

\item{maxRows}{Numeric | character. 多行时显示的最大行数}

\item{minRows}{Numeric | character. 多行时显示的最小行数}

\item{multiline}{Logical. 是否渲染为多行文本区域}

\item{name}{Character. 输入元素的名称}

\item{onChange}{值改变时的回调函数}

\item{placeholder}{Character. 输入前的提示文本}

\item{required}{Logical. 是否为必填项}

\item{rows}{Numeric | character. 多行时显示的行数}

\item{select}{Logical. 是否渲染为选择框}

\item{size}{A value equal to: 'medium', 'small', 'string'. 组件的尺寸}

\item{sx}{List of named list | logicals | named list. 允许定义系统覆盖以及其他CSS样式的系统属性}

\item{type}{Character. 输入元素的类型}

\item{value}{Logical | numeric | character | named list | unnamed list. 输入元素的值}

\item{variant}{A value equal to: 'filled', 'outlined', 'standard'. 组件的样式}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

