% Auto-generated: do not edit by hand
\name{''MuiButton}

\alias{''MuiButton}

\title{MuiButton component}

\description{

}

\usage{
''MuiButton(children=NULL, id=NULL, nClicks=NULL, variant=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. 组件型，按钮内嵌元素}

\item{id}{Character. 组件唯一id}

\item{nClicks}{Numeric. 按钮累计点击次数，用于监听按钮点击行为
默认值：`0`}

\item{variant}{A value equal to: 'text', 'outlined', 'contained', 'string'. 组件样式}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

